Add core elements #JC19-326
