+++ libs/core/guards/error.guard.ts
@@ -1,0 +1,9 @@
+import { HttpErrorDto } from '../dtos/httpError.dto';
+
+/**
+ * Check if the error is of type HttpErrorDto.
+ * @param error Some error.
+ */
+export function isHttpErrorDto(error: unknown): error is HttpErrorDto {
+  return (error as HttpErrorDto).detail !== undefined;
+}
